requirements.txt
channels==3.0.4
channels-redis==3.3.1
Django==3.2.11
djangorestframework==3.13.1